refactor(health): split 821-line runtime_snapshot.go by concern (core→207)
Pure same-package relocation, no call-site or behavior changes:

- runtime_snapshot.go (207): snapshot data model + CollectLifecycleRuntimeSnapshot entry
- runtime_snapshot_collect.go: filesystem/env/sqlite collection helpers
- runtime_snapshot_compose.go: docker-compose spec parsing + normalization
- runtime_snapshot_assess.go: drift assessment + comparison helpers

22 funcs in, 22 funcs out (lossless). Verified by go build, go vet,
the health package tests, scoped golangci-lint, goimports.
